Add PHP 5.4 as `RECOMMENDED_VERSION`
Older PHP versions are no longer actively supported by this addon
and as such will be restricted to a new `frozen` update tier,
which will only receive new fixes for issues deemed critical.

It is highly recommended to upgrade to a newer PHP version
to have access to all the latest features.